Fix gapless playback for MP3
This correctly implements gapless playback. Previous commits already fixed applying the encoder delay and padding, this patch also adds the decoder delay (which is apparently 529 for libMAD. It also fixes an issue where the first mpeg frame with the Xing and Lame headers was inadvertently send to the decoder, which added another 1152 frames of silence to the audio.